rename(Manual): 메뉴얼 팝업 위치 수정
