+++ src/components/ManualPopup/ManualPopup.tsx
